Change classes names for more consistency
