--- Terminal.Gui/Drivers/FakeDriver/IFakeConsoleInput.cs
@@ -1,4 +1,0 @@
-﻿namespace Terminal.Gui.Drivers;
-
-internal interface IFakeConsoleInput : IConsoleInput<ConsoleKeyInfo>
-{ }
