+++ Terminal.Gui/Drivers/FakeDriver/IFakeInput.cs
@@ -1,0 +1,4 @@
+﻿namespace Terminal.Gui.Drivers;
+
+internal interface IFakeInput : IConsoleInput<ConsoleKeyInfo>
+{ }
