--- Tests/TerminalGuiFluentTesting/FakeDriver/FakeFakeConsoleInput.cs
@@ -1,4 +1,0 @@
-﻿namespace TerminalGuiFluentTesting;
-
-internal class FakeFakeConsoleInput (CancellationToken hardStopToken) : FakeInput<ConsoleKeyInfo> (hardStopToken), IFakeConsoleInput
-{ }
