+++ Tests/TerminalGuiFluentTesting/FakeDriver/FakeFakeInput.cs
@@ -1,0 +1,4 @@
+﻿namespace TerminalGuiFluentTesting;
+
+internal class FakeFakeInput (CancellationToken hardStopToken) : FakeConsoleInput<ConsoleKeyInfo> (hardStopToken), IFakeInput
+{ }
