+++ Tests/TerminalGuiFluentTesting/FakeDriver/FakeUnixInput.cs
@@ -1,5 +1,5 @@
 ﻿
 namespace TerminalGuiFluentTesting;
 
-internal class FakeUnixInput (CancellationToken hardStopToken) : FakeInput<char> (hardStopToken), IUnixInput
+internal class FakeUnixInput (CancellationToken hardStopToken) : FakeConsoleInput<char> (hardStopToken), IUnixInput
 { }
