+++ Tests/TerminalGuiFluentTesting/FakeFakeComponentFactory.cs
@@ -1,7 +1,7 @@
 ﻿namespace TerminalGuiFluentTesting;
 
 #pragma warning disable CS1591
-internal class FakeFakeComponentFactory (FakeFakeConsoleInput fakeInput, FakeOutput output, ConsoleSizeMonitor fakeSizeMonitor) : FakeComponentFactory
+internal class FakeFakeComponentFactory (FakeFakeInput fakeInput, FakeOutput output, ConsoleSizeMonitor fakeSizeMonitor) : FakeComponentFactory
 {
     /// <inheritdoc/>
     public override IConsoleInput<ConsoleKeyInfo> CreateInput () { return fakeInput; }
